+++ lib/src/commands/convert_command.dart
@@ -63,7 +63,7 @@ class ConvertCommand extends Command<int> {
         needBuildRunner: false,
         useFvm: false,
         useShorebird: false,
-        shorebirdAutoConfirm: true,
+        shorebirdNoConfirm: true,
         keystorePath: 'android/key.properties',
       );
     }